+++ R/T1EpSceptical.R
@@ -1,6 +1,6 @@
 #' @export
 .T1EpSceptical_ <- function(level, c,
-                            alternative = c("one.sided", "two.sided", "greater", "less"),
+                            alternative = c("one.sided", "two.sided"),
                             type = c("golden", "nominal", "controlled")) {  
 
     stopifnot(is.numeric(level),
@@ -57,24 +57,8 @@
             }
         }
     }
+  
     if (alternative == "one.sided") {
-        ## if c = 1 compute analytically
-        if (c == 1) {
-            t1err <- 1 - stats::pnorm(q = 2*zas)
-            return(t1err)
-        } else { ## if c != 1 use numerical integration
-
-                                        # define function to integrate over zo from zas to Infty
-            intFun <- function(zo) {
-                K <- zo^2/zas^2
-                ## compute minimal zr to achieve replication success given zo and level
-                zrmin <- zas*sqrt(1 + c/(K - 1))
-                ## compute integrand: P(zr >= zrmin)*dnorm(zo)
-                (1 - stats::pnorm(q = zrmin))*stats::dnorm(x = zo)
-            }
-        }
-    }
-    if (alternative == "greater") {
         ## if c = 1 compute analytically
         if (c == 1) {
             t1err <- (1 - stats::pnorm(q = 2*zas))/2
@@ -91,37 +75,18 @@
             }
         }
     }
-    if (alternative == "less") {
-        ## if c = 1 compute analytically
-        if (c == 1) {
-            t1err <- stats::pnorm(q = 2*zas)/2
-            return(t1err)
-        } else { ## if c != 1 use numerical integration
 
-            ## define function to integrate over zo from zas to Infty
-            intFun <- function(zo) {
-                K <- zo^2/zas^2
-                ## compute maximal zr to achieve replication success given zo and level
-                zrmax <- zas*sqrt(1 + c/(K - 1))
-                ## compute integrand: P(zr <= zrmax)*dnorm(zo)
-                stats::pnorm(q = zrmax)*stats::dnorm(x = zo)
-            }
-        }
-    }
-    if (alternative %in% c("one.sided", "two.sided")) {
+    if (alternative == "two.sided") {
         ## the integral is symmetric around zero for "one.sided" and "two.sided"
         ## so we can multiply the integral from zas to Infty by 2
         t1err <- 2*stats::integrate(f = intFun, lower = zas, upper = Inf)$value
         return(t1err)
     }
-    if (alternative == "greater") {
+    if (alternative == "one.sided") {
         t1err <- stats::integrate(f = intFun, lower = zas, upper = Inf)$value
         return(t1err)
     }
-    if (alternative == "less") {
-        t1err <- stats::integrate(f = intFun, lower = -Inf, upper = zas)$value
-        return(t1err)
-    }
+
     return(t1err)
 }
 
@@ -138,15 +103,7 @@
 #' size of the replication study to the sample size of the
 #' original study.
 #' @param alternative Specifies if \code{level}
-#' is "two.sided" or one.sided ("one.sided", "greater", or "less").
-#' If "one.sided", the type-I error rate is computed based on a one-sided assessment
-#' of replication success in the direction of the original effect estimate.
-#' If "two.sided", the type-I error rate is computed based
-#' on a two-sided assessment of replication success regardless of the direction
-#' of the original and replication effect estimate.
-#' If "greater" or "less",  the type-I error rate is
-#' computed based on a one-sided assessment of replication success
-#' in the pre-specified direction of the original and replication effect estimate.
+#' is "two.sided" or "one.sided".
 #' @param type Type of recalibration. Recalibration type can be either "golden"
 #' (default), "nominal" (no recalibration), or "controlled".
 #' @return The overall type-I error rate.
@@ -172,7 +129,7 @@
 #' types <- c("nominal", "golden", "controlled")
 #' c <- seq(0.2, 5, by = 0.05)
 #' t1 <- sapply(X = types, FUN = function(t) {
-#'   T1EpSceptical(type = t, c = c, alternative = "greater", level = 0.025)
+#'   T1EpSceptical(type = t, c = c, alternative = "one.sided", level = 0.025)
 #' })
 #' matplot(x = c, y = t1*100, type = "l", lty = 1, lwd = 2, las = 1, log = "x",
 #'         xlab = bquote(italic(c)), ylab = "Type-I error (%)", xlim = c(0.2, 5))