+++ NEWS.md
@@ -6,6 +6,11 @@
 - Update links throughout the entire repository following maintainer and ownership change
 - Implemented `"controlled"` level 
 - Added one-sided p-values in `RProjects'
+- Alternatives `"greater"` and `"less"` in functions  `"T1EpSceptical"` and 
+`"PPpSceptical"` were removed. 
+Alternative `"one.sided"` now returns the T1E rate (respectively project power) 
+assuming both alternatives are one-sided in the same direction. 
+- `"levelEquivalent"` and `"unirootAll"` were removed
 
 
 # ReplicationSuccess 1.2